2010-06-19 16:24 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * include/hbdefs.h     + Adjusted HB_SIZE_MAX value for Win64 after recent changes.
  * src/vm/hashfunc.c
  * src/vm/hvm.c
  * src/debug/dbgentry.c
  * src/common/hbwin.c
  * src/common/hbfsapi.c
  * src/common/hbstr.c
  * src/rtl/lennum.c
  * src/rtl/hbzlibgz.c
  * src/rtl/hbtoken.c
  * src/rtl/hbmd5.c
  * src/rtl/hbregex.c
  * src/rtl/hbzlib.c
  * src/rtl/gtpca/gtpca.c
  * contrib/hbct/screen2.c
  * contrib/hbct/token2.c
  * contrib/hbct/charsprd.c
  * contrib/hbct/charlist.c
  * contrib/hbct/charsort.c
  * contrib/hbct/maxline.c
  * contrib/hbct/wordrepl.c
  * contrib/hbct/charrepl.c
  * contrib/hbct/charswap.c
  * contrib/hbct/pos1.c
  * contrib/hbct/token1.c
  * contrib/hbct/charmirr.c
  * contrib/hbct/relation.c
  * contrib/hbct/atnum.c
  * contrib/hbct/count.c
  * contrib/hbct/ctstrfil.c
  * contrib/hbct/posdiff.c
  * contrib/hbct/addascii.c
  * contrib/hbct/pos2.c
  * contrib/hbct/numat.c
  * contrib/hbct/atrepl.c
  * contrib/hbct/print.c
  * contrib/xhb/txtline.c
  * contrib/hbgt/strright.c
  * contrib/hbgt/chrtotal.c
  * contrib/hbgt/strcount.c
  * contrib/hbgt/ascposgt.c
  * contrib/hbgt/strleft.c
  * contrib/hbgt/chrcount.c
  * contrib/hbgt/strcspn.c
  * contrib/hbgt/atdiff.c
  * contrib/hbbz2/hbbz2.c
  * contrib/hbmemio/memio.c
  * contrib/hbclipsm/num.c
  * contrib/hbclipsm/date.c
  * contrib/hbmisc/stringsx.c
  * contrib/hbmisc/spd.c
    + Using new size API.
      NOTE: This will be a longer process as there many current variations
            to pass sizes (*nint(), *ni(), *nl()) and there is no warning
            issued in several of these.
            Please also note that not all updates files became fully clean.
            I will post a list of remaining warnings.
    + Type cleanup:
        int -> HB_ISIZ
    + Added casts at component boundaries.
    % Deleted unnecessary '( HB_SIZE )' casts.
    + ULONG_MAX -> HB_SIZE_MAX
    * 'hb_retl( 0 )' -> 'hb_retl( HB_FALSE )'
    * 'hb_retl( 1 )' -> 'hb_retl( HB_TRUE )'
    ! CSETSAFETY() fixed to return and accept logical parameters
      instead of returning numeric, checking for logical but retrieving
      numeric.

  * src/common/hbwin.c
    * DWORD -> int to be in sync with Windows API declarations.

  ; NOTE: I left out RDDs and VM from the areas I scanned, please
          modify them. F.e. we will need a hb_vmPushSize().

  ; Peer review (by as much eyeballs as possible) is appreciated.

+++ harbour/contrib/hbbz2/hbbz2.c
@@ -89,7 +89,7 @@ static int hb_bz2Compress( const char * szSrc, HB_SIZE nSrc,
    stream.avail_in  = ( unsigned int ) nSrc;
 
    stream.next_out  = szDst;
-   stream.avail_out = *pnDst;
+   stream.avail_out = ( unsigned int ) *pnDst;
 
    stream.bzalloc   = hb_bz2Alloc;
    stream.bzfree    = hb_bz2Free;
@@ -177,7 +177,7 @@ static int hb_bz2Uncompress( const char * szSrc, HB_SIZE nSrc,
    stream.avail_in  = ( unsigned int ) nSrc;
 
    stream.next_out  = szDst;
-   stream.avail_out = *pnDst;
+   stream.avail_out = ( unsigned int ) *pnDst;
 
    stream.bzalloc   = hb_bz2Alloc;
    stream.bzfree    = hb_bz2Free;
